refactor plugin loading helpers
